app-wide refactoring of id type from Long to Integer; auth controller tests

+++ recat/src/main/java/com/code/recat/author/AuthorRepository.java
@@ -1,5 +1,6 @@
 package com.code.recat.author;
 
+import jakarta.transaction.Transactional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AuthorRepository extends JpaRepository<Author, Long> {
+public interface AuthorRepository extends JpaRepository<Author, Integer> {
     Page<Author> getAllByOrderByName(Pageable pageable);
 
     @Query("SELECT a FROM Author a WHERE LOWER(a.name) = LOWER(:authorName)")
