+++ recat/src/main/java/com/code/recat/comment/CommentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
 
-    Page<Comment> findAllByOrderByCreatedAtDesc(Pageable pageable, Long bookId);
+    Page<Comment> findAllByOrderByCreatedAtDesc(Pageable pageable, Integer bookId);
 }
